Moved common components to a separate folder.
