+++ src/api/util/helpers.js
@@ -28,3 +28,11 @@ export const prepareData = values => {
 
   return values;
 };
+
+export const arrayToString = value => {
+  if (value && Array.isArray(value)) {
+    return value.join('; ');
+  }
+
+  return value;
+};
