+++ src/components/Layout/BasicMenu.js
@@ -27,9 +27,12 @@ class BasicMenu extends Component {
           theme="dark"
           inlineCollapsed={this.props.collapsed}
         >
-          <SubMenu key="organization"
-                   title={<span><Icon type="share-alt"/><span><FormattedMessage id="menu.organization"
-                                                                                defaultMessage="Organizations"/></span></span>}>
+          <SubMenu key="organization" title={
+            <div>
+              <Icon type="share-alt"/>
+              <FormattedMessage id="menu.organization" defaultMessage="Organizations"/>
+            </div>
+          }>
             <Menu.Item key="/organization/search">
               <FormattedMessage id="menu.search" defaultMessage="Search"/>
             </Menu.Item>
@@ -42,10 +45,17 @@ class BasicMenu extends Component {
             <Menu.Item key="/organization/nonPublishing">
               <FormattedMessage id="menu.organization_nonPublishing" defaultMessage="Non publishing organizations"/>
             </Menu.Item>
+            <Menu.Item key="/organization/create">
+              <FormattedMessage id="menu.createOrganization" defaultMessage="Create new organization"/>
+            </Menu.Item>
           </SubMenu>
 
-          <SubMenu key="dataset" title={<span><Icon type="table"/><span><FormattedMessage id="menu.dataset"
-                                                                                          defaultMessage="Datasets"/></span></span>}>
+          <SubMenu key="dataset" title={
+            <div>
+              <Icon type="table"/>
+              <FormattedMessage id="menu.dataset" defaultMessage="Datasets"/>
+            </div>
+          }>
             <Menu.Item key="/dataset/search">
               <FormattedMessage id="menu.search" defaultMessage="Search"/>
             </Menu.Item>
@@ -60,8 +70,12 @@ class BasicMenu extends Component {
             </Menu.Item>
           </SubMenu>
 
-          <SubMenu key="installation" title={<span><Icon type="hdd"/><span><FormattedMessage id="menu.installation"
-                                                                                             defaultMessage="Installations"/></span></span>}>
+          <SubMenu key="installation" title={
+            <div>
+              <Icon type="hdd"/>
+              <FormattedMessage id="menu.installation" defaultMessage="Installations"/>
+            </div>
+          }>
             <Menu.Item key="/installation/search">
               <FormattedMessage id="menu.search" defaultMessage="Search"/>
             </Menu.Item>
@@ -73,8 +87,12 @@ class BasicMenu extends Component {
             </Menu.Item>
           </SubMenu>
 
-          <SubMenu key="grbio" title={<span><Icon type="api"/><span><FormattedMessage id="menu.grbio"
-                                                                                      defaultMessage="GRBIO"/></span></span>}>
+          <SubMenu key="grbio" title={
+            <div>
+              <Icon type="api"/>
+              <FormattedMessage id="menu.grbio" defaultMessage="GRBIO"/>
+            </div>
+          }>
             <Menu.Item key="/grbio/collection/search">
               <FormattedMessage id="menu.collection" defaultMessage="Collections"/>
             </Menu.Item>
